refactor : based on previous review
