--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,21 +1,0 @@
-
-## 요구사항
-
-### 기본
-- [x] 기본 항목 1
-- [ ] 기본 항목 2
-
-### 심화
-- [ ] 심화 항목 1
-- [ ] 심화 항목 2
-
-## 주요 변경사항
-- 
-- 
-
-## 스크린샷
-![image](이미지url)
-
-## 멘토에게
-- 셀프 코드 리뷰를 통해 질문 이어가겠습니다.
--
