+++ sprint1/ElectronicProduct.js
