+++ sprint1/main.js
@@ -2,21 +2,9 @@ import Product from "./Product.js";
 import ElectronicProduct from "./ElectronicProduct.js";
 import Article from "./Article.js";
 
-import {
-  getProductList,
-  getProduct,
-  createProduct,
-  patchProduct,
-  deleteProduct,
-} from "./ProductService.js";
+import ps from "./ProductService.js";
 
-import {
-  getArticle,
-  createArticle,
-  patchArticle,
-  deleteArticle,
-  getArticleList,
-} from "./ArticleService.js";
+import as from "./ArticleService.js";
 
 // Product
 console.log("=== Product ===");
@@ -24,7 +12,7 @@ console.log("=== Product ===");
 // createProduct
 console.log("--- createProduct + getProduct ---");
 
-const productId = await createProduct(
+const productId = await ps.createProduct(
   ["https://www.examples.com"],
   ["전자제품"],
   2000,
@@ -33,13 +21,13 @@ const productId = await createProduct(
 );
 
 // getProduct
-const productData = await getProduct(productId);
+const productData = await ps.getProduct(productId);
 console.log(productData);
 
 // patchProduct
 console.log("--- patchProduct ---");
 
-const patchedProductId = await patchProduct(
+const patchedProductId = await ps.patchProduct(
   productId,
   ["https://www.patched.com"],
   ["전자제품"],
@@ -47,18 +35,18 @@ const patchedProductId = await patchProduct(
   "패치했습니다.",
   "아이패드 패치"
 );
-const patchedProduct = await getProduct(patchedProductId);
+const patchedProduct = await ps.getProduct(patchedProductId);
 console.log(patchedProduct);
 
 // deleteProduct
 console.log("--- deleteProduct ---");
 
-const deletedProductId = await deleteProduct(patchedProductId);
+const deletedProductId = await ps.deleteProduct(patchedProductId);
 console.log(deletedProductId);
 
 // getProductList
 console.log("--- getProductList ---");
-const productList = await getProductList(1, 2, "패드");
+const productList = await ps.getProductList(1, 2, "패드");
 console.log(productList);
 // Product의 리스트로 변형
 
@@ -89,33 +77,33 @@ console.log(products);
 console.log("=== Article ===");
 // createArticle + getArticle
 console.log("--- createArticle + getArticle ---");
-const articleId = await createArticle(
+const articleId = await as.createArticle(
   "샘플 기사",
   "샘플 기사엔 샘플이 있습니다.",
   "https://www.example.com"
 );
-const articleData = await getArticle(articleId);
+const articleData = await as.getArticle(articleId);
 console.log(articleData);
 
 // patchArticle
 console.log("--- patchArticle ---");
-const patchedArticleId = await patchArticle(
+const patchedArticleId = await as.patchArticle(
   articleId,
   "패치 기사",
   "패치되었습니다.",
   "https://www.example.com"
 );
-const patchedArticle = await getArticle(patchedArticleId);
+const patchedArticle = await as.getArticle(patchedArticleId);
 console.log(patchedArticle);
 
 // deleteArticle
 console.log("--- deleteArticle ---");
-const deletedArticleId = await deleteArticle(patchedArticleId);
+const deletedArticleId = await as.deleteArticle(patchedArticleId);
 console.log(deletedArticleId);
 
 // getArticleList
 console.log("--- getArticleList ---");
-const articleList = await getArticleList(1, 2, "기사");
+const articleList = await as.getArticleList(1, 2, "기사");
 console.log(articleList);
 
 // Article의 리스트로 변경
